Restructure boot process around `script.ltx` for flexibility
