+++ gamedata/configs/mod_script_amx.ltx
@@ -1,0 +1,3 @@
+![common]
+script 						= xr/compiler, xr/lua, amx, _G, _g_patches, xr/classes, xr/processes, dxml_core
+>class_registrators 			= amx/registrator.register
