--- gamedata/configs/mod_script_dxml.ltx
@@ -1,3 +1,0 @@
-![common]
->script 						= _g_patches, dxml_core
->class_registrators 			= class_registrator_modded_exes.register
