+++ gamedata/scripts/amx/init.lua
@@ -1,0 +1,8 @@
+-- Anomaly Modded eXes Entrypoint
+-- Extends base X-Ray script functionality
+
+--- Setup unlocalizer data model
+require(_PACKAGE .. "/unlocalize")
+
+--- Patch xr/lua compiler with modded exes extensions
+require(_PACKAGE .. "/lua")
