+++ gamedata/scripts/init.lua
@@ -109,8 +109,5 @@ end
 
 package.loaders = { _LOADERS.init }
 
---- Initialize environment via the boot module
+--- Hand control to the boot module
 require("boot")
-
--- Pass control to xray entrypoint
-require("xr")
