--- gamedata/scripts/scam/init.lua
@@ -1,9 +1,0 @@
---- Script Compilers And Macros
---- Script preprocess dispatch machinery
-
--- Setup compiler
-local compiler = require("scam/compiler")
-
-return {
-   compiler = compiler
-}
